+++ src/basic/hash_mem.h
@@ -14,12 +14,12 @@ static inline size_t ht_mem(size_t bktsize, size_t nbkts, size_t nbits) {
 }
 
 // Returns capacity of a hash table that holds at least nkmers
-size_t hash_table_cap(size_t nkmers, uint64_t *num_bkts_ptr, uint8_t *bkt_size_ptr);
+size_t hash_table_cap(uint64_t nkmers, uint64_t *num_bkts_ptr, uint8_t *bkt_size_ptr);
 
 // Returns memory required to hold nkmers
-size_t hash_table_mem(size_t nkmers, size_t entrybits, size_t *nkmers_ptr);
+size_t hash_table_mem(uint64_t nkmers, size_t entrybits, uint64_t *nkmers_ptr);
 
 // Returns memory used for hashtable no more than some memory limit
-size_t hash_table_mem_limit(size_t memlimit, size_t entrybits, size_t *nkmers_ptr);
+size_t hash_table_mem_limit(size_t memlimit, size_t entrybits, uint64_t *nkmers_ptr);
 
 #endif /* HASH_MEM_H_ */
